Move BackgroundMessageHandlers tests out of background.test.js, and add new tests for new method
